BUG: Restore removed slicer.util._executePythonModule
`slicer.util._executePythonModule` function has been widely used in many extensions. Recently this function was moved to packaging, and since then these extensions no longer work.
The name started with an underscore because it was originally intended to be an internal function, but then it turned out to be essential, and we started using it in extensions.
It should be actually a public utility function.

This commits restore `slicer.util._executePythonModule` to make extensions work again, but makes it deprecated.
From now on, `slicer.util.executePythonModule` function should be used instead.
