+++ Base/Python/slicer/packaging.py
@@ -20,9 +20,7 @@ from __future__ import annotations
 import importlib
 import importlib.metadata
 import logging
-import os
 import shlex
-import shutil
 import sys
 import tomllib
 from pathlib import Path
@@ -34,7 +32,7 @@ from packaging.requirements import InvalidRequirement, Requirement
 from packaging.utils import canonicalize_name
 
 from slicer.i18n import tr as _
-from slicer.util import launchConsoleProcess, logProcessOutput
+from slicer.util import executePythonModule
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -42,57 +40,9 @@ if TYPE_CHECKING:
     import qt
 
 
-def _executePythonModule(
-    module: str,
-    args: list[str],
-    blocking: bool = True,
-    logCallback: Callable[[str], None] | None = None,
-    completedCallback: Callable[[int], None] | None = None,
-) -> None:
-    """Execute a Python module as a script in Slicer's Python environment.
-
-    Internally python -m is called with the module name and additional arguments.
-
-    :param module: Python module name to execute (e.g., "pip")
-    :param args: command-line arguments to pass to the module
-    :param blocking: If True (default), block until completion. If False, return immediately.
-    :param logCallback: When blocking=False, called with each line of output.
-    :param completedCallback: When blocking=False, called when process completes.
-    :raises RuntimeError: if PythonSlicer executable not found
-    :raises CalledProcessError: in blocking mode, if process fails
-    """
-    # Determine pythonSlicerExecutablePath
-    try:
-        from slicer import app  # noqa: F401
-
-        # If we get to this line then import from "app" is succeeded,
-        # which means that we run this function from Slicer Python interpreter.
-        # PythonSlicer is added to PATH environment variable in Slicer
-        # therefore shutil.which will be able to find it.
-        pythonSlicerExecutablePath = shutil.which("PythonSlicer")
-        if not pythonSlicerExecutablePath:
-            raise RuntimeError("PythonSlicer executable not found")
-    except ImportError:
-        # Running from console
-        pythonSlicerExecutablePath = os.path.dirname(sys.executable) + "/PythonSlicer"
-        if os.name == "nt":
-            pythonSlicerExecutablePath += ".exe"
-
-    commandLine = [pythonSlicerExecutablePath, "-m", module, *args]
-
-    if blocking:
-        proc = launchConsoleProcess(commandLine, useStartupEnvironment=False)
-        logProcessOutput(proc, logCallback=logCallback)
-    else:
-        launchConsoleProcess(
-            commandLine,
-            useStartupEnvironment=False,
-            blocking=False,
-            logCallback=logCallback,
-            completedCallback=completedCallback,
-        )
-
-
+# Tests can monkeypatch _executePythonModule to perform mock actions instead of actually executing
+# the Python module by using slicer.packaging._executePythonModule.
+_executePythonModule = executePythonModule
 def load_requirements(path: str | Path) -> list[Requirement]:
     """Load requirements from a requirements.txt file.
